Refactor delta tower behaviors into helper module
